рефакторит компоненты после ревью кода
