+++ project/src/components/pages/favorites/favorites-place-card/favorites-place-card.jsx
@@ -6,9 +6,9 @@ function PlaceCard({ price, previewImage, apartmentType, rating, title, id }) {
   return (
     <article className="favorites__card place-card">
       <div className="favorites__image-wrapper place-card__image-wrapper">
-        <a href="#">
+        <Link to={`/offer/${id}`}>
           <img className="place-card__image" src={previewImage} width="150" height="110" alt="Place image" />
-        </a>
+        </Link>
       </div>
       <div className="favorites__card-info place-card__info">
         <div className="place-card__price-wrapper">
