+++ project/src/components/pages/main/place-card/place-card.jsx
@@ -6,9 +6,9 @@ function PlaceCard({ price, previewImage, title, apartmentType, rating, onMouseO
   return (
     <article className="cities__place-card place-card" onMouseOver={onMouseOver}>
       <div className="cities__image-wrapper place-card__image-wrapper">
-        <a href="#">
+        <Link to={`/offer/${id}`}>
           <img className="place-card__image" src={previewImage} width="260" height="200" alt={title} />
-        </a>
+        </Link>
       </div>
       <div className="place-card__info">
         <div className="place-card__price-wrapper">
